Refactor subscriptions app: Rename and update models
This commit refactors the subscriptions app by renaming the `SubscriptionRequest` model to `JobSeekerApplication` and introducing a new `TraineeApplication` model. The models were updated with new fields to better reflect the application data being collected. The old migration file was removed and a new one was created to reflect these changes.